chore(biome): add ultracite and update project structure and enhance logging functionality
- Migrated configuration to Biome by updating `biome.json` with new linter rules and extending from `ultracite`.
- Updated `package.json` to include `ultracite` as a dependency and adjusted scripts for compatibility with Biome.
- Refactored logging implementation by creating a new `build-log-message.ts` and `create-logger.ts` for improved log message formatting and handling.
- Added new tests for logger and transport functionalities, ensuring comprehensive coverage for logging behavior and server operations.
- Renamed and reorganized several files for better clarity and consistency in the codebase.
