+++ package.json
@@ -1,85 +1,86 @@
 {
-	"name": "logixlysia",
-	"version": "4.1.2",
-	"description": "🦊 Logixlysia is a logger for Elysia",
-	"type": "module",
-	"module": "src/index.ts",
-	"main": "src/index.ts",
-	"author": "PunGrumpy",
-	"publisher": "PunGrumpy",
-	"maintainers": ["PunGrumpy"],
-	"contributors": [
-		{
-			"name": "n0ky4",
-			"url": "https://github.com/n0ky4"
-		}
-	],
-	"publishConfig": {
-		"access": "public"
-	},
-	"license": "MIT",
-	"scripts": {
-		"lint": "biome check .",
-		"format": "biome check --write .",
-		"test": "bun test --timeout 5000",
-		"test:ci": "bun test --timeout 5000 --coverage --coverage-reporter=lcov",
-		"publish": "npm publish",
-		"dev": "bun run --watch example/basic.ts",
-		"prepare": "husky",
-		"test:coverage": "bun test --timeout 5000 --coverage",
-		"clean": "git clean -fdx && rm -rf node_modules && rm -rf logs"
-	},
-	"os": ["darwin", "linux", "win32"],
-	"badges": [
-		{
-			"name": "npm",
-			"url": "https://img.shields.io/npm/v/logixlysia.svg",
-			"description": "npm version",
-			"href": "https://www.npmjs.com/package/logixlysia"
-		},
-		{
-			"name": "npm",
-			"url": "https://img.shields.io/npm/dt/logixlysia.svg",
-			"description": "npm downloads",
-			"href": "https://www.npmjs.com/package/logixlysia"
-		},
-		{
-			"name": "GitHub issues",
-			"url": "https://img.shields.io/github/issues/PunGrumpy/logixlysia.svg",
-			"description": "GitHub issues",
-			"href": "https://github.com/PunGrumpy/logixlysia/issues"
-		}
-	],
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/PunGrumpy/logixlysia.git"
-	},
-	"bugs": {
-		"url": "https://github.com/PunGrumpy/logixlysia/issues"
-	},
-	"homepage": "https://github.com/PunGrumpy/logixlysia#readme",
-	"keywords": [
-		"web",
-		"logging",
-		"logger",
-		"elysia",
-		"elysiajs",
-		"logixlysia",
-		"middleware"
-	],
-	"dependencies": {
-		"chalk": "^5.3.0",
-		"elysia": "^1.1.23"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "^1.9.4",
-		"@elysiajs/eden": "^1.1.3",
-		"@elysiajs/swagger": "^1.2.0",
-		"bun-types": "^1.1.33",
-		"globals": "^16.0.0",
-		"husky": "^9.1.6"
-	},
-	"peerDependencies": {
-		"typescript": "^5.2.2"
-	}
+  "name": "logixlysia",
+  "version": "4.1.2",
+  "description": "🦊 Logixlysia is a logger for Elysia",
+  "type": "module",
+  "module": "src/index.ts",
+  "main": "src/index.ts",
+  "author": "PunGrumpy",
+  "publisher": "PunGrumpy",
+  "maintainers": ["PunGrumpy"],
+  "contributors": [
+    {
+      "name": "n0ky4",
+      "url": "https://github.com/n0ky4"
+    }
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT",
+  "scripts": {
+    "lint": "ultracite lint",
+    "format": "ultracite format",
+    "test": "bun test --timeout 5000",
+    "test:ci": "bun test --timeout 5000 --coverage --coverage-reporter=lcov",
+    "publish": "npm publish",
+    "dev": "bun run --watch example/basic.ts",
+    "prepare": "husky",
+    "test:coverage": "bun test --timeout 5000 --coverage",
+    "clean": "git clean -fdx && rm -rf node_modules && rm -rf logs"
+  },
+  "os": ["darwin", "linux", "win32"],
+  "badges": [
+    {
+      "name": "npm",
+      "url": "https://img.shields.io/npm/v/logixlysia.svg",
+      "description": "npm version",
+      "href": "https://www.npmjs.com/package/logixlysia"
+    },
+    {
+      "name": "npm",
+      "url": "https://img.shields.io/npm/dt/logixlysia.svg",
+      "description": "npm downloads",
+      "href": "https://www.npmjs.com/package/logixlysia"
+    },
+    {
+      "name": "GitHub issues",
+      "url": "https://img.shields.io/github/issues/PunGrumpy/logixlysia.svg",
+      "description": "GitHub issues",
+      "href": "https://github.com/PunGrumpy/logixlysia/issues"
+    }
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/PunGrumpy/logixlysia.git"
+  },
+  "bugs": {
+    "url": "https://github.com/PunGrumpy/logixlysia/issues"
+  },
+  "homepage": "https://github.com/PunGrumpy/logixlysia#readme",
+  "keywords": [
+    "web",
+    "logging",
+    "logger",
+    "elysia",
+    "elysiajs",
+    "logixlysia",
+    "middleware"
+  ],
+  "dependencies": {
+    "chalk": "^5.3.0",
+    "elysia": "^1.1.23"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "^1.9.4",
+    "@elysiajs/eden": "^1.1.3",
+    "@elysiajs/swagger": "^1.2.0",
+    "bun-types": "^1.1.33",
+    "globals": "^16.0.0",
+    "husky": "^9.1.6",
+    "ultracite": "^4.2.1"
+  },
+  "peerDependencies": {
+    "typescript": "^5.2.2"
+  }
 }
