+++ src/extensions/index.ts
@@ -1,1 +1,1 @@
-export { default as startServer } from './startServer'
+export { default as startServer } from './start-server'
