+++ src/helpers/method.ts
@@ -1,11 +1,11 @@
-import { HttpMethodColorMap } from './colorMapping'
+import { HttpMethodColorMap } from './color-mapping'
 
 export default function methodString(
-	method: string,
-	useColors: boolean
+  method: string,
+  useColors: boolean
 ): string {
-	const colorFunction = HttpMethodColorMap[method]
-	return useColors && colorFunction
-		? colorFunction(method.padEnd(7))
-		: method.padEnd(7)
+  const colorFunction = HttpMethodColorMap[method]
+  return useColors && colorFunction
+    ? colorFunction(method.padEnd(7))
+    : method.padEnd(7)
 }
