+++ tsconfig.json
@@ -1,25 +1,25 @@
 {
-	"$schema": "https://json.schemastore.org/tsconfig",
-	"compilerOptions": {
-		"noUncheckedIndexedAccess": true,
-		"strict": true,
-		"allowImportingTsExtensions": true,
-		"module": "ESNext",
-		"moduleResolution": "Bundler",
-		"resolveJsonModule": true,
-		"types": ["bun-types"],
-		"downlevelIteration": true,
-		"noEmit": true,
-		"allowJs": true,
-		"allowSyntheticDefaultImports": true,
-		"forceConsistentCasingInFileNames": true,
-		"jsx": "react",
-		"jsxFactory": "ElysiaJSX",
-		"jsxFragmentFactory": "ElysiaJSX.Fragment",
-		"lib": ["ESNext"],
-		"moduleDetection": "force",
-		"target": "ESNext",
-		"composite": true,
-		"skipLibCheck": true
-	}
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "noUncheckedIndexedAccess": true,
+    "strict": true,
+    "allowImportingTsExtensions": true,
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
+    "resolveJsonModule": true,
+    "types": ["bun-types"],
+    "downlevelIteration": true,
+    "noEmit": true,
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "forceConsistentCasingInFileNames": true,
+    "jsx": "react",
+    "jsxFactory": "ElysiaJSX",
+    "jsxFragmentFactory": "ElysiaJSX.Fragment",
+    "lib": ["ESNext"],
+    "moduleDetection": "force",
+    "target": "ESNext",
+    "composite": true,
+    "skipLibCheck": true
+  }
 }
